requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-common==1.1.28
azure-core==1.34.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
distro==1.9.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
langdetect==1.0.9
lxml==5.4.0
oauthlib==3.2.2
openai==1.82.0
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyMuPDF==1.26.0
pyparsing==3.2.3
python-docx==1.1.2
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
